+++ Application/Src/utils.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2024 AVSystem <avsystem@avsystem.com>
+ * Copyright 2020-2025 AVSystem <avsystem@avsystem.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,20 +46,28 @@ void get_uid(device_id_t *out_id) {
 }
 
 #ifdef USE_FW_UPDATE
-void flash_aligned_writer_new(uint64_t *batch_buf,
-                              size_t batch_buf_max_len_words,
-                              flash_aligned_writer_cb_t *writer_cb,
-                              flash_aligned_writer_t *out_writer) {
+int flash_aligned_writer_new(uint64_t *batch_buf,
+                             size_t batch_buf_max_len_words,
+                             flash_aligned_writer_cb_t *writer_cb,
+                             flash_aligned_writer_t *out_writer,
+                             size_t flash_alignment) {
     assert(batch_buf);
     assert(batch_buf_max_len_words);
     assert(writer_cb);
 
-    out_writer->batch_buf = batch_buf;
     out_writer->batch_buf_max_len_bytes =
             sizeof(*out_writer->batch_buf) * batch_buf_max_len_words;
+    if ((out_writer->batch_buf_max_len_bytes % flash_alignment) != 0) {
+        return -1;
+    }
+
+    out_writer->batch_buf = batch_buf;
     out_writer->batch_buf_len_bytes = 0;
     out_writer->write_offset_bytes = 0;
     out_writer->writer_cb = writer_cb;
+    out_writer->flash_alignment = flash_alignment;
+
+    return 0;
 }
 
 int flash_aligned_writer_write(flash_aligned_writer_t *writer,
@@ -94,9 +102,15 @@ int flash_aligned_writer_flush(flash_aligned_writer_t *writer) {
     if (writer->batch_buf_len_bytes == 0) {
         return 0;
     }
-    if (writer->batch_buf_len_bytes % sizeof(*writer->batch_buf) != 0) {
-        return -1;
+    size_t padding = writer->flash_alignment
+                     - (writer->batch_buf_len_bytes % writer->flash_alignment);
+    if (padding > 0) {
+        memset((uint8_t *) writer->batch_buf + writer->batch_buf_len_bytes,
+               0xFF, padding);
+        writer->batch_buf_len_bytes += padding;
     }
+
+    assert((writer->batch_buf_len_bytes % writer->flash_alignment) == 0);
 
     int res = writer->writer_cb(writer->batch_buf, writer->write_offset_bytes,
                                 writer->batch_buf_len_bytes);